Optimise Day 14
Instead beginning at start position every time we drop a grain, it's
faster to keep track of the positions visited in order, and start from
the penultimate position last time.

I've implemented this appraoch for part 2 in
do_part_2_by_dropping_grains, but it is not as fast as the row-by-row
approach in do_part_2.