Day 17 - get part 2 to work for test input
We now consider the relative position of the top point
in each column. Part 2 now works for the test input
and my puzzle input.

The new approach feels less robust than before - if
there are differences below the top point, it may be
possible for a shape to end up somewhere else when the
cycle repeats

For example, consider these towers:
    0123456   0123456
    █ █████   █ █████
    █         █ █
    █         █
    █         █
    █         █
    ███████   ███████

A vertical shape can end up in column 2 in the first
tower, but it is blocked in the second tower and ends
up in column 3. However, the new heuristic considers
the towers to be equivalent. This means that the
assumption that the tower height increases by the same
amount for each cycle might not be true for all possible
puzzle inputs.

    0123456   0123456
    █ █████   █ █████
    █ @       █@█
    █ @       █@
    █ @       █@
    █ @       █@
    ███████   ███████